+++ src/main/java/blackjack/domain/gamer/Gamer.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import blackjack.domain.RoundResult;
+import blackjack.domain.round.RoundResult;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import blackjack.domain.deck.Deck;
