+++ src/main/java/blackjack/dto/response/FinalResultResponseDto.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import blackjack.domain.RoundResult;
+import blackjack.domain.round.RoundResult;
 import blackjack.domain.gamer.Dealer;
 import blackjack.domain.gamer.Gamer;
 import blackjack.domain.gamer.Player;
